requirements.txt
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-vector-stores-chroma
chromadb
streamlit
python-dotenv
pypdf
sentence-transformers
tqdm
